Fix bug in TrackObject declaration.
Fix a bug where the parser, in absence of a zone specifier but in
presence of a incr one, when read the initial `in` would then expect a
ZoneValue to come next.

e.g: track "person" incr person_counter

Would fail, when the parser finds `incr`, and it reads the `in`, expects
a zone.

The fix is to just change the zone specifier from just `in` to `inzone`.
We could also change the `incr` to another synonym, but I didn't find a
good one.